requirements.txt
torch
transformers
datasets
peft
accelerate
sentencepiece
chromadb
sentence-transformers
fastapi
uvicorn
mcp
streamlit
ruff
mypy
pytest
scikit-learn
python-dotenv
requests
tqdm
llama-cpp-python
rouge-score